Replace sequencercache implementation with a hashtable and priority queue. Discard using caffien.

+++ infrastructure/src/main/java/org/corfudb/infrastructure/SequencerServer.java
@@ -155,7 +155,7 @@ public class SequencerServer extends AbstractServer {
 
         globalLogTail = Address.getMinAddress();
 
-        this.cache = new SequencerServerCache(config.getCacheSize());
+        this.cache = new SequencerServerCache((int)config.getCacheSize());
 
 
         setUpTimerNameCache();
